feat: add repositories

+++ src/services/product.service.ts
@@ -1,11 +1,18 @@
 import { Product } from '../entity';
-import { ProductRepository } from '../repositories';
+import { ErrorHandler } from '../helpers/ErrorHandler';
+import { ProductRepository, StockRepository, StoreRepository } from '../repositories';
 
 export class ProductService {
-	constructor(private productRepository: ProductRepository) {}
+	constructor(private productRepository: ProductRepository, private storeRepository: StoreRepository, private stockRepository: StockRepository) {}
 
 	async getAll(): Promise<any[]> {
-		return this.productRepository.getAll();
+		const stock = await this.stockRepository.getAll();
+		const products = await this.productRepository.getAll();
+		const productsWithStock = products.map((product) => {
+			const stockProduct = stock.find((s) => s.productId === product.id);
+			return { ...product, quantity: stockProduct?.quantity || 0 };
+		});
+		return productsWithStock;
 	}
 
 	async create(product: Product): Promise<void> {
@@ -13,6 +20,10 @@ export class ProductService {
 	}
 
 	async delete(id: string): Promise<void> {
+		const haveStoreProcess = await this.storeRepository.validateProcessByProductId(id);
+		if (haveStoreProcess) {
+			throw ErrorHandler.badRequest('Este produto está sendo utilizado em um processo de compra');
+		}
 		await this.productRepository.delete(id);
 	}
 }
